Refactored dir structure to make room for variable-specific util helpers
