+++ .gitignore
@@ -7,7 +7,5 @@ dist/*
 tests/token.ts
 tests/token.js
 src/examples/token.ts
-comp
 comps.json
-styles
 styles.json
